# Fix bug in `run_circuit_and_measure` and speed up function execution