There was a bug in the `run_circuit_and_measure` function of the SymbolicSimulator, which wasn't appearing in the tests, because positive tests weren't implemented for that function after it was integrated with our `Wavefunction` class. I fixed the bug and added the tests.

In addition, I improved the performance of the function significantly, by binding symbols' values to the circuit first, before retrieving the wavefunction. I did a simple quantitative test and this was the result:
- Run symbolic circuit then bind to wavefunction:
```python
170 ms ± 548 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
- Bind to symbolic circuit and then get wavefunction:
```python
12.9 ms ± 40.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
The test circuit was:
```python
circ = circuits.Circuit(
        [
            circuits.XX(sympy.Symbol("theta"))(0, 1),
            circuits.U3(
                sympy.Symbol("alpha"), sympy.Symbol("beta"), sympy.Symbol("gamma")
            )(2),
            circuits.XX(sympy.Symbol("mana"))(1, 2),
        ]
    )
```

This result matches our intuition since sympy calculations are relatively slower than those of numpy.